backend/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
oracledb==2.0.1

# AI
anthropic==0.39.0
httpx==0.27.2

# Data processing
pandas==2.2.0
numpy==1.26.4
pdfplumber==0.10.3
PyPDF2==3.0.1

# Cloud provider SDKs
boto3==1.34.44
azure-mgmt-compute==30.6.0
azure-identity==1.15.0
google-cloud-billing==1.12.1
google-auth==2.27.0
oci==2.119.1

# Utilities
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
tqdm==4.66.1
APScheduler==3.10.4

# Export
reportlab==4.1.0
openpyxl==3.1.2
